数据库迁移代码bug修复
